refactor: WeatherContext로 분리
