+++ src/App.tsx
@@ -15,26 +15,30 @@ import SelectDiary from './pages/analysis/SelectDiary';
 import EmotionAndQuest from './pages/analysis/EmotionAndQuest';
 import { UserProvider } from './shared/context/UserContext';
 import { PATHS } from './shared/constants/path';
+import { WeatherProvider } from './shared/components/DiaryWeather/context/WeatherContext';
 
 function App() {
   return (
     <UserProvider>
-      <BrowserRouter>
-        <Toaster position="top-center" toastOptions={{ duration: 4000 }} /> {/* react-hot-toast */}
-        <Header />
-        <Routes>
-          <Route path={PATHS.LOGIN} element={<Login />} />
-          <Route path={PATHS.REGISTER} element={<Register />} />
-          <Route path="/diary" element={<DiaryPage />} />
-          <Route path="/diary/form" element={<DiaryFormPage />} />
-          <Route path="/diary/detail" element={<DiaryDetailPage />} />
-          <Route path="/community" element={<DiaryList />} />
-          <Route path="/users" element={<UserPage />} />
-          <Route path={PATHS.MYPAGE} element={<Mypage />} />
-          <Route path="/analysis/select-diary" element={<SelectDiary />} />
-          <Route path="/analysis/emotion-and-quest" element={<EmotionAndQuest />} />
-        </Routes>
-      </BrowserRouter>
+      <WeatherProvider>
+        <BrowserRouter>
+          <Toaster position="top-center" toastOptions={{ duration: 4000 }} />{' '}
+          {/* react-hot-toast */}
+          <Header />
+          <Routes>
+            <Route path={PATHS.LOGIN} element={<Login />} />
+            <Route path={PATHS.REGISTER} element={<Register />} />
+            <Route path="/diary" element={<DiaryPage />} />
+            <Route path="/diary/form" element={<DiaryFormPage />} />
+            <Route path="/diary/detail" element={<DiaryDetailPage />} />
+            <Route path="/community" element={<DiaryList />} />
+            <Route path="/users" element={<UserPage />} />
+            <Route path={PATHS.MYPAGE} element={<Mypage />} />
+            <Route path="/analysis/select-diary" element={<SelectDiary />} />
+            <Route path="/analysis/emotion-and-quest" element={<EmotionAndQuest />} />
+          </Routes>
+        </BrowserRouter>
+      </WeatherProvider>
     </UserProvider>
   );
 }
